pyproject.toml
[project]
name = "dpi-filters"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "path>=17.1.1",
    "pillow>=12.1.1",
]
